solver: parse 64 bit ints
While the usual parse method was only capable of parsing 32 bits, not also
parse 64 bit integers.

+++ utils/ParseUtils.h
@@ -80,8 +80,8 @@ static void skipLine(B& in) {
 
 
 template<class B>
-static int parseInt(B& in) {
-    int     val = 0;
+static int64_t parseInt(B& in) {
+    int64_t     val = 0;
     bool    neg = false;
     skipWhitespace(in);
     if      (*in == '-') neg = true, ++in;
